% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMessage.R
\name{checkAFFargs}
\alias{checkAFFargs}
\title{Check the arguments for AFF (no change detection)}
\usage{
checkAFFargs(eta, functionName = "")
}
\arguments{
\item{eta}{The value of the step size in the gradient descent. 
Must be finite, in range [0,1], etc.}

\item{functionName}{The name of the original function which is being
called to check the parameters. Will help user
with debugging. Default is \code{""}, in which
case no function name will be displayed.}
}
\description{
A function which will do all the checks for AFF, check the parameters
are in the proper ranges, etc. Does not return any values.
}
\keyword{internal}
