% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMessage.R
\name{checkCUSUMMeanCDargs}
\alias{checkCUSUMMeanCDargs}
\title{Check the arguments for CUSUM mean change detector}
\usage{
checkCUSUMMeanCDargs(k, h, BL, functionName = "")
}
\arguments{
\item{k}{One of the CUSUM control parameters. Must be finite, etc.}

\item{h}{One of the CUSUM control parameters. Must be finite, etc.}

\item{BL}{The value of the burn-in length. Must be finite, above 0, etc.}

\item{functionName}{The name of the original function which is being
called to check the parameters. Will help user
with debugging. Default is \code{""}, in which
case no function name will be displayed.}
}
\description{
A function which will do all the checks for AFF, check the parameters
are in the proper ranges, etc. Does not return any values.
}
\keyword{internal}
