% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMessage.R
\name{checkFFFargs}
\alias{checkFFFargs}
\title{Check the arguments for FFF (no change detection)}
\usage{
checkFFFargs(lambda, functionName = "")
}
\arguments{
\item{lambda}{The value of the FFF. Must be finite, in range [0,1], etc.}

\item{functionName}{The name of the original function which is being
called to check the parameters. Will help user
with debugging. Default is \code{""}, in which
case no function name will be displayed.}
}
\description{
A function which will do all the checks for FFF, check the parameters
are in the proper ranges, etc. Does not return any values.
}
\keyword{internal}
