% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMessage.R
\name{checkFFFMeanCDargs}
\alias{checkFFFMeanCDargs}
\title{Check the arguments for FFF change detection initialisation}
\usage{
checkFFFMeanCDargs(alpha, lambda, BL, functionName = "")
}
\arguments{
\item{alpha}{The value of the threshold. Must be finite, in (0,1), etc.}

\item{lambda}{The value of the FFF. Must be finite, in range [0,1], etc.}

\item{BL}{The value of the burn-in length. Must be finite, above 0, etc.}

\item{functionName}{The name of the original function which is being
called to check the parameters. Will help user
with debugging. Default is \code{""}, in which
case no function name will be displayed.}
}
\value{
The truncated burn-in length (since it must be an integer).
}
\description{
A function which will do all the checks for FFFcd, check the parameters
are in the proper ranges, etc. Does not return any values.
}
\keyword{internal}
