% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMessage.R
\name{checkPrechange}
\alias{checkPrechange}
\title{Check the prechange mean, sigma and variance are appropriately defined}
\usage{
checkPrechange(prechangeMean, prechangeSigma, prechangeVar, functionName = "")
}
\arguments{
\item{prechangeMean}{The prechange mean. Should not be \code{NULL}.}

\item{prechangeSigma}{The prechange standard deviation. 
Either it or \code{prechangeVar} must not 
be \code{NULL}.}

\item{prechangeVar}{The prechange variance.
Either it or \code{prechangeSigma} must not 
be \code{NULL}.}

\item{functionName}{The name of the original function which is being
called to check the parameters. Will help user
with debugging. Default is \code{""}, in which
case no function name will be displayed.}
}
\description{
If \code{usePrechange} was set to \code{TRUE}, then \code{prechangeMean}
must not be \code{NULL}, and at least one of \code{prechangeSigma} and
\code{prechangeVar} must not be \code{NULL}.
}
\keyword{internal}
