% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMessage.R
\name{checkStream}
\alias{checkStream}
\alias{checkStream}
\title{Check the stream consists of finite numeric values}
\usage{
checkStream(x, functionName = "")

checkStream(x, functionName = "")
}
\arguments{
\item{x}{The stream}

\item{functionName}{The name of the original function which is being
called to check the parameters. Will help user
with debugging. Default is \code{""}, in which
case no function name will be displayed.}

\item{skipChecks}{A boolean}

\item{functionName}{The name of the original function which is being
called to check the parameters. Will help user
with debugging. Default is \code{""}, in which
case no function name will be displayed.}
}
\description{
Check the stream does not have any NULLs, NAs, infs, etc

Check if the skipChecks parameter is a boolean (it should be).
}
\keyword{internal}
