% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combineTwoOneSidedPvaluesR}
\alias{combineTwoOneSidedPvaluesR}
\title{Combine two p-values into a single p-value}
\usage{
combineTwoOneSidedPvaluesR(p1, p2)
}
\arguments{
\item{p1}{The first p-value.}

\item{p2}{The second p-value.}
}
\description{
Wrapper function to combine two p-values into a single p-value.
Uses a type of 'reverse Bonferroni'.
}
\keyword{internal}
