% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeTwoSidedPvalueR}
\alias{computeTwoSidedPvalueR}
\title{Compute a two-sided p-value from a quantile value}
\usage{
computeTwoSidedPvalueR(x, a, b)
}
\arguments{
\item{x}{The two-sided p-value to be converted.}

\item{a}{The lower bound, e.g. 0.}

\item{b}{The upper bound, e.g. 2.}
}
\description{
Wrapper function to compute a two-sided p-value from a statistics
or quantile value, with the aid of lower and upper bounds.
}
\keyword{internal}
