% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertPvalueToCorrectSideR}
\alias{convertPvalueToCorrectSideR}
\title{Convert a 'p-value' to the correct side.}
\usage{
convertPvalueToCorrectSideR(p)
}
\arguments{
\item{p}{The two-sided p-value to be converted.}
}
\description{
Wrapper function to convery a two-sided p-value to a one-sided
'p-value' (values close to 0 and close to 1 both become close to 0).
For example, both 0.01 and 0.99 are mapped to 0.01.
}
\keyword{internal}
