% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_computeAFFMean}
\alias{cpp_computeAFFMean}
\title{Compute the AFF mean of a vector}
\usage{
cpp_computeAFFMean(x, eta)
}
\arguments{
\item{x}{Vector of numeric values values.}

\item{eta}{Value for the step size in the gradient descent step.}
}
\description{
Given a vector \code{x} and a value \code{eta} for step size
in the stochastic gradient descent for the adaptive forgetting
factor, this returns the value of the fixed forgetting factor mean
\eqn{\bar{x}_{N, \overrightarrow{\lambda} }}, where \eqn{N} is the 
length of \code{x}. Algorithm is implemented in C++.
}
\keyword{internal}
