% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_computeFFFMean}
\alias{cpp_computeFFFMean}
\title{Compute the FFF mean of a vector}
\usage{
cpp_computeFFFMean(x, lambda)
}
\arguments{
\item{x}{Vector of numeric values values.}

\item{lambda}{Value for the fixed forgetting factor in \eqn{[0,1]}.}
}
\description{
Given a vector \code{x} and a value \code{lambda} for a fixed forgetting
factor, returns the value of the fixed forgetting factor mean
\eqn{\bar{x}_{N, \lambda}}, where \eqn{N} is the length of \code{x}.
Algorithm is implemented in C++.
}
\keyword{internal}
