% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_detectAFFMeanSingle}
\alias{cpp_detectAFFMeanSingle}
\title{Find the first changepoint in the mean using AFF}
\usage{
cpp_detectAFFMeanSingle(x, alpha, eta, BL)
}
\arguments{
\item{x}{Vector of numeric values values.}

\item{alpha}{Value for the significance threshold in \eqn{(0,1)}.}

\item{eta}{Value for the step size in \eqn{(0,1)}.}

\item{BL}{Value for the burn-in length.}
}
\value{
A list with 
\describe{
            \item{\code{tauhat}}{The index of the first changepoint found.}
         }
}
\description{
Given a vector \code{x}, a threshold \code{alpha}, a step size \code{eta}, 
and a burn-in length \code{BL}, returns a list containing the single 
changepoint. Algorithm is implemented in C++.
}
\keyword{internal}
