% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_detectFFFMeanMultiple}
\alias{cpp_detectFFFMeanMultiple}
\title{Search for multiple changepoints in the mean using FFF}
\usage{
cpp_detectFFFMeanMultiple(x, lambda, alpha, BL)
}
\arguments{
\item{x}{Vector of numeric values values.}

\item{lambda}{Value for the fixed forgetting factor in \eqn{(0,1)}.}

\item{alpha}{Value for the significance threshold in \eqn{(0,1)}.}

\item{BL}{Value for the burn-in length.}
}
\value{
A list with 
\describe{
            \item{\code{tauhat}}{A vector of the changepoints found.}
         }
}
\description{
Given a vector \code{x}, a threshold \code{alpha}, a value \code{lambda},
and a burn-in length \code{BL}, returns a list containing the changepoints.
Algorithm is implemented in C++.
}
\keyword{internal}
