% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMessage.R
\name{doesNotExist}
\alias{doesNotExist}
\title{Check if an object exists}
\usage{
doesNotExist(object)
}
\arguments{
\item{object}{The object.}
}
\description{
For a particular parameter, check if the object with that name exists
}
\keyword{internal}
