% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensureFolderDoesNotEndInSep}
\alias{ensureFolderDoesNotEndInSep}
\title{Checks that the last character of the folder is the separator}
\usage{
ensureFolderDoesNotEndInSep(folder)
}
\arguments{
\item{folder}{A string, the folder name}
}
\value{
The folder name, WITHOUT the separator, if it is present.
}
\description{
If the last character in a folder's name is the file separator, remove it.
}
\details{
The reason this is included is because the R function \code{file.path}
will not realise the last element is the platform's file separator, 
so concatenating could result in a double slash.

The reason this is included is because the R function 
         \code{file.path} will not realise the last element is 
         the platform's file separator, so concatenating could 
         result in a double slash.
}
\keyword{internal}
