% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{EWMA_stream_jumpdetect}
\alias{EWMA_stream_jumpdetect}
\title{EWMA change detection for a stream in R with}
\usage{
EWMA_stream_jumpdetect(stream, BL, params)
}
\arguments{
\item{stream}{The stream of observations.}

\item{BL}{The burn-in length, used to estimate the mean and variance.}

\item{params}{A list of parameters for the EWMA algorithm. Consists of
             \describe{
                 \item{\code{r}}{A control parameter which controls
                                 the rate of downweighting.}

                 \item{\code{L}}{A control parameter which determines 
                                 the width of the control limits.}
             }}
}
\value{
A vector of estimated changepoints.
}
\description{
Original implementation in R of EWMA change detector, now with
documentation.
}
\keyword{internal}
