% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffstream-package.R
\name{ffstream_vignette}
\alias{ffstream_vignette}
\title{ffstream vignette 0.1.6}
\usage{
ffstream_vignette()
}
\description{
This function opens the ffstream vignette.
}
\examples{
ffstream_vignette()
}
