% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{get_L_deriv}
\alias{get_L_deriv}
\title{Compute the derivative of L}
\usage{
get_L_deriv(xbar_old, xbar_old_deriv, x_new)
}
\arguments{
\item{xbar_old}{The old (previous) value of \code{xbar}.}

\item{xbar_old_deriv}{The old (previous) value of the derivative
of \code{xbar}.}

\item{x_new}{The value of the current observation.}
}
\description{
Compute the derivative of the cost function L, defined as:
\deqn{L_{N} = [\bar{x}_{N-1, \overrightarrow{\lambda}} - x_{N}}
}
\keyword{internal}
