% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{get_nextobs_fromstream}
\alias{get_nextobs_fromstream}
\title{Obtain the next observation}
\usage{
get_nextobs_fromstream(stream, streampos)
}
\arguments{
\item{stream}{The stream (vector) of observations.}

\item{streampos}{The index of the stream position.}
}
\description{
Obtain the next observation from the stream
}
\keyword{internal}
