% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{get_xbar}
\alias{get_xbar}
\title{Compute \code{xbar}}
\usage{
get_xbar(m, w)
}
\arguments{
\item{m}{The current value of \code{m}.}

\item{w}{The current value of \code{w}.}
}
\description{
Function to calculate \code{xbar} from \code{m} and \code{w}
}
\keyword{internal}
