% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{get_xbar_deriv}
\alias{get_xbar_deriv}
\title{Compute the derivative of \code{xbar}}
\usage{
get_xbar_deriv(Delta, xbar, Omega, w)
}
\arguments{
\item{Delta}{The value of \code{Delta}, the derivative of \code{m}.}

\item{xbar}{The value of the forgetting factor mean.}

\item{Omega}{The value of \code{Omega}, the derivative of \code{w}.}

\item{w}{The value of the weight used to compute \code{xbar}.}
}
\description{
Function to compute the derivative of \code{xbar}.
}
\keyword{internal}
