% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMessage.R
\name{getExistsErrorMessage}
\alias{getExistsErrorMessage}
\title{Create the error message for a certain parameter that does not exist}
\usage{
getExistsErrorMessage(paramName, value = NULL, algoName = NULL,
  functionName = "")
}
\arguments{
\item{paramName}{Name of parameter, e.g. \code{lambda}}

\item{value}{The name of the object that does not exist.}

\item{algoName}{Name of algorithm, e.g. \code{AFF change detector}}

\item{functionName}{The name of the original function which is being
called to check the parameters. Will help user
with debugging.}
}
\description{
For a particular parameter, create an error message specifying that
the parameter does not exist (e.g. \code{Na}, which is not the same
as \code{NA}).
}
\keyword{internal}
