% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMessage.R
\name{getFiniteErrorMessage}
\alias{getFiniteErrorMessage}
\title{Create the error message for a certain parameter that is not finite}
\usage{
getFiniteErrorMessage(paramName, value = NULL, algoName = NULL,
  functionName = "")
}
\arguments{
\item{paramName}{Name of parameter, e.g. \code{lambda}}

\item{value}{The value of the non-finite parameter}

\item{algoName}{Name of algorithm, e.g. \code{AFF change detector}}

\item{functionName}{The name of the original function which is being
called to check the parameters. Will help user
with debugging.}
}
\description{
For a particular parameter, create an error message specifying that
the parameter is not finite (i.e. not numeric and not NaN)
}
\keyword{internal}
