% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMessage.R
\name{getMissingErrorMessage}
\alias{getMissingErrorMessage}
\title{Create the error message for a certain parameter that is missing}
\usage{
getMissingErrorMessage(paramName, algoName = NULL, functionName = "")
}
\arguments{
\item{paramName}{Name of parameter, e.g. \code{lambda}}

\item{algoName}{Name of algorithm, e.g. \code{AFF change detector}}

\item{functionName}{The name of the original function which is being
called to check the parameters. Will help user
with debugging.}
}
\description{
For a particular parameter, create an error message specifying that
the parameter is missing.
}
\keyword{internal}
