% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{getNormalLimitsFromList}
\alias{getNormalLimitsFromList}
\title{Compute confidence limits assuming the normal distribution}
\usage{
getNormalLimitsFromList(burnvec, u_new)
}
\arguments{
\item{burnvec}{A vector of \emph{unnamed} values, consisting of:
               \describe{
                         \item{p}{First component - value of threshold.}

                         \item{mu}{Second component - value of the mean.}

                         \item{sd}{Third component - value of the standardi
                                    deviation.}

                        }}

\item{u_new}{The value of \code{u}, the quantity used to scale the 
forgetting factor variance.}
}
\description{
Gets the 100p% confidence interval limits for N(mu, stddev).
NB: stddev used, not variance!
}
\keyword{internal}
