% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{getv_from_u_and_w}
\alias{getv_from_u_and_w}
\title{Calculate \code{v}}
\usage{
getv_from_u_and_w(u, w)
}
\arguments{
\item{u}{The current value of \code{u}.}

\item{w}{The current value of \code{w}.}
}
\description{
Function to calculate \code{v} from \code{u} and \code{w}, according 
to the formula \eqn{v= w(1-u)}.
}
\keyword{internal}
