% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{inbounds}
\alias{inbounds}
\title{Check if a value is inside bounds}
\usage{
inbounds(x, low_bound, up_bound)
}
\arguments{
\item{x}{The value to check.}

\item{low_bound}{The lower bound.}

\item{up_bound}{The upper bound.}
}
\description{
Function to check that a value \code{x} is between two values
}
\keyword{internal}
