% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{initFFFMean}
\alias{initFFFMean}
\title{Initialisation of FFF}
\usage{
initFFFMean(lambda = 1)
}
\arguments{
\item{lambda}{The value of the fixed forgetting factor. 
Default is lambda=1.}
}
\description{
This function makes it simple to initalise an FFF mean object.
}
\examples{
library(Rcpp)
fff1 <- initFFFMean()     # initialises with lambda=1

fff2 <- initFFFMean(0.9)  # initialises with lambda=0.9

}
