% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isAboveBound}
\alias{isAboveBound}
\title{Check if an argument above a certain value}
\usage{
isAboveBound(x, lower)
}
\arguments{
\item{x}{The value to check.}

\item{lower}{The lower bound.}
}
\description{
Function to check that a value \code{x} is above a certain lower bound.
}
\keyword{internal}
