% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isInBounds}
\alias{isInBounds}
\title{Check if an argument is in the closed set [lower, upper]}
\usage{
isInBounds(x, lower, upper)
}
\arguments{
\item{x}{The value to check.}

\item{lower}{The lower bound.}

\item{upper}{The upper bound.}
}
\description{
Function to check that a value \code{x} is between two values.
}
\keyword{internal}
