% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isInBoundsStrictly}
\alias{isInBoundsStrictly}
\title{Check if an argument is in the open set (lower, upper)}
\usage{
isInBoundsStrictly(x, lower, upper)
}
\arguments{
\item{x}{The value to check.}

\item{lower}{The lower bound.}

\item{upper}{The upper bound.}
}
\description{
Function to check that a value \code{x} is between strictly two values.
}
\keyword{internal}
