% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{isInLimitsNormal}
\alias{isInLimitsNormal}
\title{Check if in the limits}
\usage{
isInLimitsNormal(x, limitsvec)
}
\arguments{
\item{x}{The value to check.}

\item{limitsvec}{The vector of limits (two values).}
}
\description{
Checks if a value is in normal limits
}
\keyword{internal}
