% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isInteger}
\alias{isInteger}
\title{Check if a value is an integer}
\usage{
isInteger(x)
}
\arguments{
\item{x}{The value to check.}
}
\description{
Function to check that a value \code{x} is an integer.
}
\keyword{internal}
