% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMessage.R
\name{modifyFunctionPrefix}
\alias{modifyFunctionPrefix}
\title{Modify the function name prefix}
\usage{
modifyFunctionPrefix(functionName)
}
\arguments{
\item{functionName}{The function name.}
}
\description{
Add some text to the function name so it fits in with the error message.
}
\keyword{internal}
