% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorMessage.R
\name{modifyValueStr}
\alias{modifyValueStr}
\title{Modify the string containing the value}
\usage{
modifyValueStr(value)
}
\arguments{
\item{value}{The value of the offending parameter.}
}
\description{
Create a string with the value to be used in an error message.
}
\keyword{internal}
