% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{update_Delta}
\alias{update_Delta}
\title{Update \code{Delta}}
\usage{
update_Delta(Delta_old, lambda_old, m_new)
}
\arguments{
\item{Delta_old}{The old (previous) value of \code{Delta}.}

\item{lambda_old}{The value of the forgetting factor.}

\item{m_new}{The new (current) value of \code{m}.}
}
\description{
Function to sequentially update the value of \code{Delta}
}
\keyword{internal}
