% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{update_ffmean}
\alias{update_ffmean}
\title{Update the FF mean (duplicate)}
\usage{
update_ffmean(ffmean_old, w_new, x_new)
}
\arguments{
\item{ffmean_old}{The old (previous) value of \code{ffmean}.}

\item{w_new}{The new (current) value of \code{w}.}

\item{x_new}{The current observation.}
}
\description{
Function to sequentially update the FF mean
}
\seealso{
\code{\link{update_mean}}
}
\keyword{internal}
