% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{update_lambda}
\alias{update_lambda}
\title{Update lambda}
\usage{
update_lambda(lambda_old, signchosen, alpha, deriv_new)
}
\arguments{
\item{lambda_old}{The old (previous) value of \code{lambda}.}

\item{signchosen}{Either \code{+1} or \code{-1}. Usually \code{-1}.}

\item{alpha}{The value of the step-size in the gradient descent.
In the paper this is named \eqn{\epsilon}.}

\item{deriv_new}{The value of the derivative at the current time.}
}
\description{
Update the forgetting factor \code{lambda} using the derivative of
a particular cost function
}
\keyword{internal}
