% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{update_m}
\alias{update_m}
\title{Calculate \code{m}}
\usage{
update_m(m_old, lambda_old, x_new)
}
\arguments{
\item{m_old}{The old (previous) value of \code{m}.}

\item{lambda_old}{The old (previous) value of \code{lambda}.}

\item{x_new}{The current observation.}
}
\description{
Function to calculate \code{m}
}
\keyword{internal}
