% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{update_Omega}
\alias{update_Omega}
\title{Update \code{Omega}}
\usage{
update_Omega(Omega_old, lambda_old, w_new)
}
\arguments{
\item{Omega_old}{The old (previous) value of \code{Omega}.}

\item{lambda_old}{The value of the forgetting factor.}

\item{m_new}{The new (current) value of \code{m}.}
}
\description{
Function to sequentially update the value of \code{Omega}
}
\keyword{internal}
