% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{update_streampos}
\alias{update_streampos}
\title{Update the stream position}
\usage{
update_streampos(streampos)
}
\arguments{
\item{streampos}{The index of the stream position.}
}
\description{
Updates the index keeping track of the stream position
}
\keyword{internal}
