% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{update_u}
\alias{update_u}
\title{Update \code{u}}
\usage{
update_u(u_old, w_new)
}
\arguments{
\item{u_old}{The old (previous) value of \code{u}.}

\item{w_new}{The new (current) value of \code{w}.}
}
\description{
Function to sequentially update the weight used to scale the forgetting
factor variance, \code{u}.
}
\keyword{internal}
