% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{update_var}
\alias{update_var}
\title{Update the FF variance}
\usage{
update_var(ffvar, ffmean_old, lambda, v_old, v_new, w_new, x_new)
}
\arguments{
\item{ffvar}{The new (current) value of the forgetting factor variance.}

\item{ffmean_old}{The old (previous) value of the forgetting factor mean.
In fact, the formula requires this value, rather than
the current value of the FF mean.}

\item{lambda}{The value of the forgetting factor.}

\item{v_old}{The old (previous) value of \code{v}.}

\item{v_new}{The new (current) value of \code{v}.}

\item{w_new}{The new (current) value of \code{w}.}

\item{x_new}{The new (current) observation.}
}
\description{
Function to sequentially update the forgetting factor variance
}
\keyword{internal}
