% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originalRcode.R
\name{update_w}
\alias{update_w}
\title{Update the weight}
\usage{
update_w(w_old, lambda_old)
}
\arguments{
\item{w_old}{The old (previous) value of \code{w}.}

\item{lambda_old}{The value of the forgetting factor. No real need to
add 'old' here, in the adaptive scheme, these are the
the values from the previous iteration.}
}
\description{
Function to sequentially update the weight used in the computation of
the forgetting factor mean
}
\keyword{internal}
